FLX-73: Split page into multiple comps
