+++ web/app/routes/settings/page.tsx
@@ -28,6 +28,10 @@ import {
   SelectValue,
 } from "~/components/ui/select";
 import type { SettingsData } from "~/services/settings";
+import {MailSettingsCard} from "~/routes/settings/mail-card";
+import {StorageSettingsCard} from "~/routes/settings/storage-card";
+import {ApplicationSettingsCard} from "~/routes/settings/application-card";
+import {ApiSettingsCard} from "~/routes/settings/api-card";
 
 export function meta({}: Route.MetaArgs) {
   return [
@@ -222,287 +226,6 @@ const SettingsPage = ({ loaderData }: Route.ComponentProps) => {
     );
   }
 
-  const storageDriverOptions = [
-    { value: "local", label: "Local Storage" },
-    { value: "aws", label: "AWS S3" },
-    { value: "backblaze", label: "Backblaze B2" },
-    { value: "dropbox", label: "Dropbox" },
-  ];
-
-  const mailDriverOptions = [
-    { value: "sendgrid", label: "SendGrid" },
-    { value: "mailgun", label: "Mailgun" },
-  ];
-
-  const mailgunRegionOptions = [
-    { value: "us", label: "United States" },
-    { value: "eu", label: "Europe" },
-  ];
-
-  const renderStorageConfiguration = () => {
-    const isStorageDisabled = !formData.allowStorage;
-
-    switch (formData.storageDriver) {
-      case "aws":
-        return (
-            <div className="space-y-6">
-              <div className="flex items-center gap-2 mb-4">
-                <Cloud className="h-4 w-4 text-blue-500" />
-                <h4 className="font-medium text-sm">AWS S3 Configuration</h4>
-              </div>
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <div className="space-y-3">
-                  <Label htmlFor="awsAccessKeyId" className="text-sm font-medium">
-                    Access Key ID
-                  </Label>
-                  <Input
-                      id="awsAccessKeyId"
-                      name="awsAccessKeyId"
-                      type="password"
-                      value={formData.awsAccessKeyId}
-                      onChange={(e) => handleInputChange("awsAccessKeyId", e.target.value)}
-                      placeholder="Your AWS Access Key ID"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="awsSecretAccessKey" className="text-sm font-medium">
-                    Secret Access Key
-                  </Label>
-                  <Input
-                      id="awsSecretAccessKey"
-                      name="awsSecretAccessKey"
-                      type="password"
-                      value={formData.awsSecretAccessKey}
-                      onChange={(e) => handleInputChange("awsSecretAccessKey", e.target.value)}
-                      placeholder="Your AWS Secret Access Key"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="awsRegion" className="text-sm font-medium">
-                    AWS Region
-                  </Label>
-                  <Input
-                      id="awsRegion"
-                      name="awsRegion"
-                      value={formData.awsRegion}
-                      onChange={(e) => handleInputChange("awsRegion", e.target.value)}
-                      placeholder="us-east-1"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-              </div>
-            </div>
-        );
-      case "backblaze":
-        return (
-            <div className="space-y-6">
-              <div className="flex items-center gap-2 mb-4">
-                <Cloud className="h-4 w-4 text-orange-500" />
-                <h4 className="font-medium text-sm">Backblaze B2 Configuration</h4>
-              </div>
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <div className="space-y-3">
-                  <Label htmlFor="backblazeKeyId" className="text-sm font-medium">
-                    Key ID
-                  </Label>
-                  <Input
-                      id="backblazeKeyId"
-                      name="backblazeKeyId"
-                      type="password"
-                      value={formData.backblazeKeyId}
-                      onChange={(e) => handleInputChange("backblazeKeyId", e.target.value)}
-                      placeholder="Your Backblaze Key ID"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="backblazeApplicationKey" className="text-sm font-medium">
-                    Application Key
-                  </Label>
-                  <Input
-                      id="backblazeApplicationKey"
-                      name="backblazeApplicationKey"
-                      type="password"
-                      value={formData.backblazeApplicationKey}
-                      onChange={(e) => handleInputChange("backblazeApplicationKey", e.target.value)}
-                      placeholder="Your Backblaze Application Key"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-              </div>
-            </div>
-        );
-      case "dropbox":
-        return (
-            <div className="space-y-6">
-              <div className="flex items-center gap-2 mb-4">
-                <Cloud className="h-4 w-4 text-blue-600" />
-                <h4 className="font-medium text-sm">Dropbox Configuration</h4>
-              </div>
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <div className="space-y-3">
-                  <Label htmlFor="dropboxAccessToken" className="text-sm font-medium">
-                    Access Token
-                  </Label>
-                  <Input
-                      id="dropboxAccessToken"
-                      name="dropboxAccessToken"
-                      type="password"
-                      value={formData.dropboxAccessToken}
-                      onChange={(e) => handleInputChange("dropboxAccessToken", e.target.value)}
-                      placeholder="Your Dropbox Access Token"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="dropboxAppKey" className="text-sm font-medium">
-                    App Key
-                  </Label>
-                  <Input
-                      id="dropboxAppKey"
-                      name="dropboxAppKey"
-                      type="password"
-                      value={formData.dropboxAppKey}
-                      onChange={(e) => handleInputChange("dropboxAppKey", e.target.value)}
-                      placeholder="Your Dropbox App Key"
-                      className="mt-2"
-                      disabled={isStorageDisabled}
-                  />
-                </div>
-              </div>
-            </div>
-        );
-      default:
-        return null;
-    }
-  };
-
-  const renderMailConfiguration = () => {
-    switch (formData.mailDriver) {
-      case "sendgrid":
-        return (
-            <div className="space-y-6">
-              <div className="flex items-center gap-2 mb-4">
-                <Mail className="h-4 w-4 text-blue-500" />
-                <h4 className="font-medium text-sm">SendGrid Configuration</h4>
-              </div>
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <div className="space-y-3">
-                  <Label htmlFor="sendgridApiKey" className="text-sm font-medium">
-                    API Key
-                  </Label>
-                  <Input
-                      id="sendgridApiKey"
-                      name="sendgridApiKey"
-                      type="password"
-                      value={formData.sendgridApiKey}
-                      onChange={(e) => handleInputChange("sendgridApiKey", e.target.value)}
-                      placeholder="Your SendGrid API Key"
-                      className="mt-2"
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="sendgridEmailSource" className="text-sm font-medium">
-                    Email Source
-                  </Label>
-                  <Input
-                      id="sendgridEmailSource"
-                      name="sendgridEmailSource"
-                      type="email"
-                      value={formData.sendgridEmailSource}
-                      onChange={(e) => handleInputChange("sendgridEmailSource", e.target.value)}
-                      placeholder="noreply@yourapp.com"
-                      className="mt-2"
-                  />
-                </div>
-              </div>
-            </div>
-        );
-      case "mailgun":
-        return (
-            <div className="space-y-6">
-              <div className="flex items-center gap-2 mb-4">
-                <Mail className="h-4 w-4 text-red-500" />
-                <h4 className="font-medium text-sm">Mailgun Configuration</h4>
-              </div>
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <div className="space-y-3">
-                  <Label htmlFor="mailgunApiKey" className="text-sm font-medium">
-                    API Key
-                  </Label>
-                  <Input
-                      id="mailgunApiKey"
-                      name="mailgunApiKey"
-                      type="password"
-                      value={formData.mailgunApiKey}
-                      onChange={(e) => handleInputChange("mailgunApiKey", e.target.value)}
-                      placeholder="Your Mailgun API Key"
-                      className="mt-2"
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="mailgunEmailSource" className="text-sm font-medium">
-                    Email Source
-                  </Label>
-                  <Input
-                      id="mailgunEmailSource"
-                      name="mailgunEmailSource"
-                      type="email"
-                      value={formData.mailgunEmailSource}
-                      onChange={(e) => handleInputChange("mailgunEmailSource", e.target.value)}
-                      placeholder="noreply@yourapp.com"
-                      className="mt-2"
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="mailgunDomain" className="text-sm font-medium">
-                    Domain
-                  </Label>
-                  <Input
-                      id="mailgunDomain"
-                      name="mailgunDomain"
-                      value={formData.mailgunDomain}
-                      onChange={(e) => handleInputChange("mailgunDomain", e.target.value)}
-                      placeholder="mg.yourapp.com"
-                      className="mt-2"
-                  />
-                </div>
-                <div className="space-y-3">
-                  <Label htmlFor="mailgunRegion" className="text-sm font-medium">
-                    Region
-                  </Label>
-                  <Select
-                      value={formData.mailgunRegion}
-                      onValueChange={(value) => handleInputChange("mailgunRegion", value)}
-                  >
-                    <SelectTrigger className="mt-2">
-                      <SelectValue placeholder="Select region" />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {mailgunRegionOptions.map(option => (
-                          <SelectItem key={option.value} value={option.value}>
-                            {option.label}
-                          </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </div>
-              </div>
-            </div>
-        );
-      default:
-        return null;
-    }
-  };
-
   return (
       <div className="min-h-screen bg-gray-50">
         <AppHeader title="Settings">
@@ -538,328 +261,23 @@ const SettingsPage = ({ loaderData }: Route.ComponentProps) => {
             {/* Application Settings & API Settings */}
             <div className="grid grid-cols-1 xl:grid-cols-2 gap-8">
               {/* Application Settings */}
-              <Card className="h-fit">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <Settings className="h-5 w-5 text-blue-500" />
-                    <CardTitle>Application Settings</CardTitle>
-                  </div>
-                  <CardDescription>
-                    Configure your application's basic settings
-                  </CardDescription>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="space-y-3">
-                    <Label htmlFor="appTitle" className="text-sm font-medium">
-                      App Title
-                    </Label>
-                    <Input
-                        id="appTitle"
-                        name="appTitle"
-                        value={formData.appTitle}
-                        onChange={(e) => handleInputChange("appTitle", e.target.value)}
-                        placeholder="Your App Name"
-                        className="mt-2"
-                    />
-                  </div>
-                  <div className="space-y-3">
-                    <Label htmlFor="appUrl" className="text-sm font-medium">
-                      App URL
-                    </Label>
-                    <Input
-                        id="appUrl"
-                        name="appUrl"
-                        type="url"
-                        value={formData.appUrl}
-                        onChange={(e) => handleInputChange("appUrl", e.target.value)}
-                        placeholder="https://your-app.com"
-                        className="mt-2"
-                    />
-                  </div>
-                  <div className="space-y-3">
-                    <Label htmlFor="jwtSecret" className="text-sm font-medium">
-                      JWT Secret
-                    </Label>
-                    <Input
-                        id="jwtSecret"
-                        name="jwtSecret"
-                        type="password"
-                        value={formData.jwtSecret}
-                        onChange={(e) => handleInputChange("jwtSecret", e.target.value)}
-                        placeholder="Your JWT secret key"
-                        className="mt-2"
-                    />
-                  </div>
-                  <div className="space-y-3">
-                    <Label htmlFor="maxProjectsPerOrg" className="text-sm font-medium">
-                      Max Projects per Organization
-                    </Label>
-                    <Input
-                        id="maxProjectsPerOrg"
-                        name="maxProjectsPerOrg"
-                        type="number"
-                        value={formData.maxProjectsPerOrg}
-                        onChange={(e) => handleInputChange("maxProjectsPerOrg", parseInt(e.target.value))}
-                        min="1"
-                        max="1000"
-                        className="mt-2"
-                    />
-                  </div>
-                  <div className="pt-4 border-t">
-                    <div className="flex items-center justify-between py-2">
-                      <Label htmlFor="allowRegistrations" className="text-sm font-medium">
-                        Allow Registrations
-                      </Label>
-                      <Switch
-                          id="allowRegistrations"
-                          name="allowRegistrations"
-                          checked={formData.allowRegistrations}
-                          onCheckedChange={(checked) => handleInputChange("allowRegistrations", checked)}
-                      />
-                    </div>
-                    <div className="flex items-center justify-between py-2">
-                      <Label htmlFor="allowProjects" className="text-sm font-medium">
-                        Allow Projects
-                      </Label>
-                      <Switch
-                          id="allowProjects"
-                          name="allowProjects"
-                          checked={formData.allowProjects}
-                          onCheckedChange={(checked) => handleInputChange("allowProjects", checked)}
-                      />
-                    </div>
-                    <div className="flex items-center justify-between py-2">
-                      <Label htmlFor="allowForms" className="text-sm font-medium">
-                        Allow Forms
-                      </Label>
-                      <Switch
-                          id="allowForms"
-                          name="allowForms"
-                          checked={formData.allowForms}
-                          onCheckedChange={(checked) => handleInputChange("allowForms", checked)}
-                      />
-                    </div>
-                    <div className="flex items-center justify-between py-2">
-                      <Label htmlFor="allowBackups" className="text-sm font-medium">
-                        Allow Backups
-                      </Label>
-                      <Switch
-                          id="allowBackups"
-                          name="allowBackups"
-                          checked={formData.allowBackups}
-                          onCheckedChange={(checked) => handleInputChange("allowBackups", checked)}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
+              <ApplicationSettingsCard formData={formData} onInputChange={handleInputChange}></ApplicationSettingsCard>
 
               {/* Storage Settings */}
-              <Card className="h-fit">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <Database className="h-5 w-5 text-purple-500" />
-                    <CardTitle>Storage Settings</CardTitle>
-                  </div>
-                  <CardDescription>
-                    Configure storage options, limits, and cloud provider settings
-                  </CardDescription>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="flex items-center justify-between py-2">
-                    <Label htmlFor="allowStorage" className="text-sm font-medium">
-                      Allow Storage
-                    </Label>
-                    <Switch
-                        id="allowStorage"
-                        name="allowStorage"
-                        checked={formData.allowStorage}
-                        onCheckedChange={(checked) => handleInputChange("allowStorage", checked)}
-                    />
-                  </div>
-
-                  <div className="space-y-6 pt-4 border-t">
-                    <div className="space-y-3">
-                      <Label htmlFor="storageDriver" className="text-sm font-medium">
-                        Storage Driver
-                      </Label>
-                      <Select
-                          value={formData.storageDriver}
-                          onValueChange={(value) => handleInputChange("storageDriver", value)}
-                          disabled={!formData.allowStorage}
-                      >
-                        <SelectTrigger className="mt-2">
-                          <SelectValue placeholder="Select storage driver" />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {storageDriverOptions.map(option => (
-                              <SelectItem key={option.value} value={option.value}>
-                                {option.label}
-                              </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-                    <div className="space-y-3">
-                      <Label htmlFor="storageMaxContainers" className="text-sm font-medium">
-                        Max Storage Containers
-                      </Label>
-                      <Input
-                          id="storageMaxContainers"
-                          name="storageMaxContainers"
-                          type="number"
-                          value={formData.storageMaxContainers}
-                          onChange={(e) => handleInputChange("storageMaxContainers", parseInt(e.target.value))}
-                          min="1"
-                          className="mt-2"
-                          disabled={!formData.allowStorage}
-                      />
-                    </div>
-                    <div className="space-y-3">
-                      <Label htmlFor="storageMaxFileSizeInKB" className="text-sm font-medium">
-                        Max File Size (KB)
-                      </Label>
-                      <Input
-                          id="storageMaxFileSizeInKB"
-                          name="storageMaxFileSizeInKB"
-                          type="number"
-                          value={formData.storageMaxFileSizeInKB}
-                          onChange={(e) => handleInputChange("storageMaxFileSizeInKB", parseInt(e.target.value))}
-                          min="1"
-                          className="mt-2"
-                          disabled={!formData.allowStorage}
-                      />
-                    </div>
-                    <div className="space-y-3">
-                      <Label htmlFor="storageAllowedMimes" className="text-sm font-medium">
-                        Allowed MIME Types
-                      </Label>
-                      <Textarea
-                          id="storageAllowedMimes"
-                          name="storageAllowedMimes"
-                          value={formData.storageAllowedMimes}
-                          onChange={(e) => handleInputChange("storageAllowedMimes", e.target.value)}
-                          placeholder="image/jpeg,image/png,application/pdf"
-                          rows={4}
-                          className="mt-2"
-                          disabled={!formData.allowStorage}
-                      />
-                    </div>
-                  </div>
-
-                  {/* Cloud Storage Configuration */}
-                  {formData.storageDriver !== "local" && (
-                      <div className="pt-6 border-t">
-                        {renderStorageConfiguration()}
-                      </div>
-                  )}
-                </CardContent>
-              </Card>
+              <StorageSettingsCard formData={formData} onInputChange={handleInputChange} ></StorageSettingsCard>
             </div>
 
             {/* Storage Settings & Mail Settings */}
             <div className="grid grid-cols-1 xl:grid-cols-2 gap-8">
               {/* API Settings */}
-              <Card className="h-fit">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <Settings className="h-5 w-5 text-orange-500" />
-                    <CardTitle>API Settings</CardTitle>
-                  </div>
-                  <CardDescription>
-                    Configure API throttling and rate limiting
-                  </CardDescription>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="flex items-center justify-between py-2">
-                    <Label htmlFor="allowApiThrottle" className="text-sm font-medium">
-                      Enable API Throttling
-                    </Label>
-                    <Switch
-                        id="allowApiThrottle"
-                        name="allowApiThrottle"
-                        checked={formData.allowApiThrottle}
-                        onCheckedChange={(checked) => handleInputChange("allowApiThrottle", checked)}
-                    />
-                  </div>
-                  <div className="space-y-6 pt-4 border-t">
-                    <div className="space-y-3">
-                      <Label htmlFor="apiThrottleLimit" className="text-sm font-medium">
-                        Throttle Limit
-                      </Label>
-                      <Input
-                          id="apiThrottleLimit"
-                          name="apiThrottleLimit"
-                          type="number"
-                          value={formData.apiThrottleLimit}
-                          onChange={(e) => handleInputChange("apiThrottleLimit", parseInt(e.target.value))}
-                          min="1"
-                          className="mt-2"
-                          disabled={!formData.allowApiThrottle}
-                      />
-                    </div>
-                    <div className="space-y-3">
-                      <Label htmlFor="apiThrottleInterval" className="text-sm font-medium">
-                        Throttle Interval (seconds)
-                      </Label>
-                      <Input
-                          id="apiThrottleInterval"
-                          name="apiThrottleInterval"
-                          type="number"
-                          value={formData.apiThrottleInterval}
-                          onChange={(e) => handleInputChange("apiThrottleInterval", parseInt(e.target.value))}
-                          min="1"
-                          className="mt-2"
-                          disabled={!formData.allowApiThrottle}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
+              <ApiSettingsCard formData={formData} onInputChange={handleInputChange}></ApiSettingsCard>
 
               {/* Mail Settings */}
-              <Card className="h-fit">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <Mail className="h-5 w-5 text-blue-500" />
-                    <CardTitle>Mail Settings</CardTitle>
-                  </div>
-                  <CardDescription>
-                    Configure email service provider and settings
-                  </CardDescription>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="space-y-3">
-                    <Label htmlFor="mailDriver" className="text-sm font-medium">
-                      Mail Driver
-                    </Label>
-                    <Select
-                        value={formData.mailDriver}
-                        onValueChange={(value) => handleInputChange("mailDriver", value)}
-                    >
-                      <SelectTrigger className="mt-2">
-                        <SelectValue placeholder="Select mail driver" />
-                      </SelectTrigger>
-                      <SelectContent>
-                        {mailDriverOptions.map(option => (
-                            <SelectItem key={option.value} value={option.value}>
-                              {option.label}
-                            </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  {/* Mail Provider Configuration */}
-                  <div className="pt-4 border-t">
-                    {renderMailConfiguration()}
-                  </div>
-                </CardContent>
-              </Card>
+              <MailSettingsCard formData={formData} onInputChange={handleInputChange}></MailSettingsCard>
             </div>
 
             {/* Save Button */}
-            <div className="flex justify-end pt-8 border-t bg-white p-6 rounded-lg shadow-sm">
+            <div className="flex justify-end">
               <Button
                   type="submit"
                   disabled={!isDirty || isSaving}
